102 Extracts WfoProductBlockKeyValueRow component
