+++ packages/orchestrator-ui-components/src/components/WfoSubscription/WfoSubscriptionProductBlock/index.ts
@@ -1,1 +1,2 @@
 export * from './WfoSubscriptionProductBlock';
+export * from './WfoProductBlockKeyValueRow';
